End seed and start of relations

+++ database/migrations/2020_03_31_005503_create_bookings_table.php
@@ -20,8 +20,8 @@ class CreateBookingsTable extends Migration
             $table->dateTime('end_date');
             $table->integer('duration_stay');
 
-            $table->unsignedInteger('id_group');
-            $table->unsignedInteger('id_booking_state');
+            $table->unsignedBigInteger('id_group');
+            $table->unsignedBigInteger('id_booking_state');
 
             $table->timestamps();
